refactor(srv): split agent_session.rs into submodules
