+++ .changesets/1782972226-refactor-srv-split-agent-session-into-submodules-aclb.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(srv): split agent_session into submodules
